+++ code/src/cljs/sixsq/nuvla/ui/edge_detail/views.cljs
@@ -197,10 +197,12 @@
         releases      (subscribe [::edge-subs/nuvlabox-releases-options])
         close-fn      #(reset! show? false)
         form-data     (r/atom nil)
+        force-restart (r/atom false)
         project       (-> @status :installation-parameters :project-name)
         working-dir   (-> @status :installation-parameters :working-dir)
         config-files  (-> @status :installation-parameters :config-files)
         environment   (-> @status :installation-parameters :environment)
+        nb-version    (get @status :nuvlabox-engine-version nil)
         on-change-fn  #(swap! form-data assoc :nuvlabox-release %)
         on-success-fn close-fn
         on-error-fn   close-fn
@@ -212,9 +214,11 @@
     (swap! form-data assoc :working-dir working-dir)
     (swap! form-data assoc :config-files (str/join "\n" config-files))
     (swap! form-data assoc :environment (str/join "\n" environment))
+    (swap! form-data assoc :force-restart false)
+    (when nb-version
+      (swap! form-data assoc :current-version nb-version))
     (fn [{:keys [id] :as _resource} _operation show? title icon button-text]
       (let [correct-nb?    (= (:parent @status) id)
-            nb-version     (get @status :nuvlabox-engine-version "")
             target-version (->> @releases
                                 (some #(when (= (:value %) (:nuvlabox-release @form-data)) %))
                                 :key)]
@@ -262,6 +266,18 @@
           [uix/Accordion
            [:<>
             [ui/Form
+             [ui/FormField
+              [:label
+               "Force Restart"]
+              [ui/Radio {:toggle    true
+                         :checked   @force-restart
+                         :label     (if (:force-restart @form-data)
+                                      (@tr [:nuvlabox-update-force-restart])
+                                      (@tr [:nuvlabox-update-no-force-restart]))
+                         :on-change #(do
+                                       (swap! force-restart not)
+                                       (swap! form-data assoc :force-restart @force-restart))}]]
+
              [ui/FormInput {:label         (str/capitalize (@tr [:project]))
                             :placeholder   "nuvlabox"
                             :required      true