+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+  - NuvlaBox update action has new option to force restart during the update operation
+
 ## [2.23.0] - 2021-10-13
 
 ### Changed